+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 Note: Can be used using nvuillam/mega-linter@insiders in your mega-linter.yml file
 
+- Core
+  - Implement architecture for Mega-Linter plugins
+
 - Reports
   - Better icons for Console, GitHub Comment and Text reporters: ✅ ❌